refactor: fix button-radio style and cleanup
