+++ src/components/forms/index.ts
@@ -3,4 +3,4 @@ export * from './TextArea';
 export * from './Checkbox';
 export * from './context';
 export * from './Select';
-export * from './RadioGroup';
+export * from './radio-group';
