+++ src/components/forms/radio-group/ClassicRadioItem.tsx
@@ -54,6 +54,7 @@ const classicStyles = {
       cursor: 'pointer',
       lineHeight: 1,
       paddingLeft: 5,
+      fontSize: '12px',
       ':hover': {
         '& ~ button': {
           borderColor: disabled ? 'grey' : enabledColor,
